De-duplicate test script
Use the same test script for `cedar-example-use-case` and
`cedar-policy-language-in-action`.

This also fixes a bug in the old test script for
`cedar-example-use-cases`. Due to how it used a sub-shell, it didn't
actually exit when it the policies it expected to validated failed to
validate. Due to a second bug, a change in text printed to stdout by the
CLI on validation success was causing the test to fail (but not exit).